Configure Uglifier to remove all comments
Uglifier is preserving copyright comments which, in many cases, produces a
non optimal result (if you are using a lot of 3rd party libraries, then
the result will probably have a lot of licence comments).

Use the following option during construction in order to remove all
comments: :output => { :comments => :none }.

TODO: this could be controlled by an optional parameter in the Guardfile.